+++ packages/swap/src/swappers/evm/evm-swapper.ts
@@ -6,6 +6,10 @@ import type { ExecuteArgs, Swapper } from '@/types'
 
 import { getExecuteData } from './execute-data'
 
+/**
+ * @description Instead of using this function directly, run token swaps as transactions with the [@rarimo/nft-checkout]() package.
+ * For more information about swaps, see [Swapping tokens](https://docs.rarimo.com/overview/swapping-tokens) in the Rarimo documentation.
+ */
 export const createEVMSwapper = (provider: IProvider): Swapper => {
   const bridger = createBridger(createEVMBridger, provider)
   const isInitialized = ref(false)